Les types des variables latitude et longitude ont été changé en double

+++ app/src/main/java/com/carpool/activity/ConsultationOffreActivity.java
@@ -127,11 +127,11 @@ package com.carpool.activity;
                     e1.printStackTrace();
                 }
 
-                int lattDep = trajetResultat.getPositionDepart().getLatitude();
-                int longDep = trajetResultat.getPositionDepart().getLongitude();
+                double lattDep = trajetResultat.getPositionDepart().getLatitude();
+                double longDep = trajetResultat.getPositionDepart().getLongitude();
 
-                int lattArr = trajetResultat.getPositionArrive().getLatitude();
-                int longArr = trajetResultat.getPositionArrive().getLongitude();
+                double lattArr = trajetResultat.getPositionArrive().getLatitude();
+                double longArr = trajetResultat.getPositionArrive().getLongitude();
 
 
                 SimpleDateFormat sdf = new SimpleDateFormat("yyyy-MM-dd");
